Integration of ML driven flutter tags
